221223 - Overwrite existing directory and complete code for python main.py file
When given an existing project name when creating a new project, user is given
the choise to overwrite the existing project or not. If yes, all project
files and subdirectories are deleted before creating the project.

Completed function write_main_python in project_tools.lua: most common
imports for ev3dev2 are included.